+++ tests/integration_test/services_test.py
@@ -35721,6 +35721,7 @@ def test_get_bot_settings():
     assert actual['data'] == {'is_billed': False, 'chat_token_expiry': 30,
                               'data_generation_limit_per_day': 3,
                               'data_importer_limit_per_day': 5,
+                              'enable_voice': False,
                               'force_import': False,
                               'ignore_utterances': False,
                               'llm_settings': {'enable_faq': False, 'provider': 'openai'},
@@ -35834,6 +35835,7 @@ def test_update_analytics_settings():
     assert actual['data'] == {'is_billed': False, 'chat_token_expiry': 30,
                               'data_generation_limit_per_day': 3,
                               'data_importer_limit_per_day': 5,
+                              'enable_voice': False,
                               'force_import': False,
                               'ignore_utterances': False,
                               'llm_settings': {'enable_faq': False, 'provider': 'openai'},
